+++ realm/realm-library/src/main/java/io/realm/RealmQuery.java
@@ -23,6 +23,7 @@ import java.util.Locale;
 
 import javax.annotation.Nullable;
 
+import io.realm.annotations.Beta;
 import io.realm.annotations.Required;
 import io.realm.internal.OsResults;
 import io.realm.internal.OsList;
@@ -33,6 +34,7 @@ import io.realm.internal.SortDescriptor;
 import io.realm.internal.Table;
 import io.realm.internal.TableQuery;
 import io.realm.internal.fields.FieldDescriptor;
+import io.realm.log.RealmLog;
 
 
 /**
@@ -63,6 +65,8 @@ public class RealmQuery<E> {
     private String className;
     private final boolean forValues;
     private final OsList osList;
+    private SortDescriptor sortDescriptor;
+    private SortDescriptor distinctDescriptor;
 
     private static final String TYPE_MISMATCH = "Field '%s': type mismatch - %s expected.";
     private static final String EMPTY_VALUES = "Non-empty 'values' must be provided.";
@@ -1810,7 +1814,7 @@ public class RealmQuery<E> {
     public RealmResults<E> findAll() {
         realm.checkIfValid();
 
-        return createRealmResults(query, null, null, true);
+        return createRealmResults(query, sortDescriptor, distinctDescriptor, true);
     }
 
     /**
@@ -1824,7 +1828,7 @@ public class RealmQuery<E> {
         realm.checkIfValid();
 
         realm.sharedRealm.capabilities.checkCanDeliverNotification(ASYNC_QUERY_WRONG_THREAD_MESSAGE);
-        return createRealmResults(query, null, null, false);
+        return createRealmResults(query, sortDescriptor, distinctDescriptor, false);
     }
 
     /**
@@ -1841,9 +1845,9 @@ public class RealmQuery<E> {
      * {@link RealmObject} or a child {@link RealmList}.
      */
     @SuppressWarnings("unchecked")
+    @Deprecated
     public RealmResults<E> findAllSorted(String fieldName, Sort sortOrder) {
         realm.checkIfValid();
-
         SortDescriptor sortDescriptor = SortDescriptor.getInstanceForSort(getSchemaConnector(), query.getTable(), fieldName, sortOrder);
         return createRealmResults(query, sortDescriptor, null, true);
     }
@@ -1857,6 +1861,7 @@ public class RealmQuery<E> {
      * @throws java.lang.IllegalArgumentException if field name does not exist or it belongs to a child
      * {@link RealmObject} or a child {@link RealmList}.
      */
+    @Deprecated
     public RealmResults<E> findAllSortedAsync(final String fieldName, final Sort sortOrder) {
         realm.checkIfValid();
 
@@ -1865,6 +1870,129 @@ public class RealmQuery<E> {
         return createRealmResults(query, sortDescriptor, null, false);
     }
 
+    /**
+     * // FIXME: I think we can sort on children now?
+     * Sort the query result by the specific field name in ascending order. This predicate can be applied multiple
+     * times, but only the last specified sorting order will be used.
+     * <p>
+     * Sorting is currently limited to character sets in 'Latin Basic', 'Latin Supplement', 'Latin Extended A',
+     * 'Latin Extended B' (UTF-8 range 0-591). For other character sets, sorting will have no effect.
+     *
+     * @param fieldName the field name to sort by.
+     * @throws java.lang.IllegalArgumentException if the field name does not exist or it belongs to a child
+     * {@link RealmObject} or a child {@link RealmList}.
+     */
+    public RealmQuery<E> sort(String fieldName) {
+        realm.checkIfValid();
+        return sort(fieldName, Sort.ASCENDING);
+    }
+
+    /**
+     * // FIXME: Sorting on children?
+     * Sort the query result by the specified field name and order. This predicate can be applied multiple
+     * times, but only the last specified sorting order will be used.
+     * <p>
+     * Sorting is currently limited to character sets in 'Latin Basic', 'Latin Supplement', 'Latin Extended A',
+     * 'Latin Extended B' (UTF-8 range 0-591). For other character sets, sorting will have no effect.
+     *
+     * @param fieldName the field name to sort by.
+     * @param sortOrder how to sort the results.
+     * @throws java.lang.IllegalArgumentException if field name does not exist or it belongs to a child
+     * {@link RealmObject} or a child {@link RealmList}.
+     */
+    public RealmQuery<E> sort(String fieldName, Sort sortOrder) {
+        realm.checkIfValid();
+        return sort(new String[] { fieldName}, new Sort[] { sortOrder});
+    }
+
+    /**
+     * // FIXME: Sorting on children?
+     * Sort the query result by the specific field names in the provided order. {@code fieldName2} is only used
+     * in case of equal values in {@code fieldName1}. This predicate can be applied multiple
+     * times, but only the last specified sorting order will be used.
+     * <p>
+     * Sorting is currently limited to character sets in 'Latin Basic', 'Latin Supplement', 'Latin Extended A',
+     * 'Latin Extended B' (UTF-8 range 0-591). For other character sets, sorting will have no effect.
+     *
+     * @param fieldName1 first field name
+     * @param sortOrder1 sort order for first field
+     * @param fieldName2 second field name
+     * @param sortOrder2 sort order for second field
+     * @throws java.lang.IllegalArgumentException if a field name does not exist or it belongs to a child
+     * {@link RealmObject} or a child {@link RealmList}.
+     */
+    public RealmQuery<E> sort(String fieldName1, Sort sortOrder1, String fieldName2, Sort sortOrder2) {
+        realm.checkIfValid();
+        return sort(new String[] { fieldName1, fieldName2 }, new Sort[] { sortOrder1, sortOrder2 });
+    }
+
+    /**
+     * // FIXME: Sorting on children?
+     * Sort the query result by the specific field names in the provided order. Later fields will only be used
+     * if the previous field values are equal. This predicate can be applied multiple times, but only the last
+     * specified sorting order will be used.
+     * <p>
+     * Sorting is currently limited to character sets in 'Latin Basic', 'Latin Supplement', 'Latin Extended A',
+     * 'Latin Extended B' (UTF-8 range 0-591). For other character sets, sorting will have no effect.
+     *
+     * @param fieldNames an array of field names to sort by.
+     * @param sortOrders how to sort the field names.
+     * @throws java.lang.IllegalArgumentException if one of the field names does not exist or it belongs to a child
+     * {@link RealmObject} or a child {@link RealmList}.
+     */
+    public RealmQuery<E> sort(String[] fieldNames, Sort[] sortOrders) {
+        realm.checkIfValid();
+        if (sortDescriptor != null) {
+            RealmLog.debug("Overriding existing sort order");
+        }
+        sortDescriptor = SortDescriptor.getInstanceForSort(getSchemaConnector(), query.getTable(), fieldNames, sortOrders);
+        return this;
+    }
+
+    /**
+     * BETA API: Will be renamed to {@code distinct} in next major release.
+     *
+     * Selects a distinct set of objects of a specific class. If the result is sorted, the first object will be
+     * returned in case of multiple occurrences, otherwise it is undefined which object is returned. This
+     * predicate can be applied multiple times, but only the last applied will be used.
+     * <p>
+     * Adding {@link io.realm.annotations.Index} to the corresponding field will make this operation much faster.
+     *
+     * @param fieldName the field name.
+     * @throws IllegalArgumentException if a field is {@code null}, does not exist, is an unsupported type, or points
+     * to linked fields.
+     */
+    @Beta
+    public RealmQuery<E> distinctValues(String fieldName) {
+        realm.checkIfValid();
+        distinctDescriptor = SortDescriptor.getInstanceForDistinct(getSchemaConnector(), query.getTable(), fieldName);
+        return this;
+    }
+
+    /**
+     * BETA API: Will be renamed to {@code distinct} in next major release.
+     *
+     * Selects a distinct set of objects of a specific class. When multiple distinct fields are
+     * given, all unique combinations of values in the fields will be returned. In case of multiple
+     * matches, it is undefined which object is returned. Unless the result is sorted, then the
+     * first object will be returned.
+     *
+     * @param firstFieldName first field name to use when finding distinct objects.
+     * @param remainingFieldNames remaining field names when determining all unique combinations of field values.
+     * @throws IllegalArgumentException if field names is empty or {@code null}, does not exist,
+     * is an unsupported type, or points to a linked field.
+     */
+    @Beta
+    public RealmQuery<E> distinctValues(String firstFieldName, String... remainingFieldNames) {
+        realm.checkIfValid();
+
+        String[] fieldNames = new String[1 + remainingFieldNames.length];
+
+        fieldNames[0] = firstFieldName;
+        System.arraycopy(remainingFieldNames, 0, fieldNames, 1, remainingFieldNames.length);
+        distinctDescriptor = SortDescriptor.getInstanceForDistinct(getSchemaConnector(), table, fieldNames);
+        return this;
+    }
 
     /**
      * Finds all objects that fulfill the query conditions and sorted by specific field name in ascending order.
@@ -1878,6 +2006,7 @@ public class RealmQuery<E> {
      * @throws java.lang.IllegalArgumentException if the field name does not exist or it belongs to a child
      * {@link RealmObject} or a child {@link RealmList}.
      */
+    @Deprecated
     public RealmResults<E> findAllSorted(String fieldName) {
         return findAllSorted(fieldName, Sort.ASCENDING);
     }
@@ -1891,6 +2020,7 @@ public class RealmQuery<E> {
      * @throws java.lang.IllegalArgumentException if the field name does not exist or it belongs to a child
      * {@link RealmObject} or a child {@link RealmList}.
      */
+    @Deprecated
     public RealmResults<E> findAllSortedAsync(String fieldName) {
         return findAllSortedAsync(fieldName, Sort.ASCENDING);
     }
@@ -1908,6 +2038,7 @@ public class RealmQuery<E> {
      * @throws java.lang.IllegalArgumentException if one of the field names does not exist or it belongs to a child
      * {@link RealmObject} or a child {@link RealmList}.
      */
+    @Deprecated
     public RealmResults<E> findAllSorted(String[] fieldNames, Sort[] sortOrders) {
         realm.checkIfValid();
 
@@ -1930,6 +2061,7 @@ public class RealmQuery<E> {
      * {@link RealmObject} or a child {@link RealmList}.
      * @see io.realm.RealmResults
      */
+    @Deprecated
     public RealmResults<E> findAllSortedAsync(String[] fieldNames, final Sort[] sortOrders) {
         realm.checkIfValid();
 
@@ -1953,6 +2085,7 @@ public class RealmQuery<E> {
      * @throws java.lang.IllegalArgumentException if a field name does not exist or it belongs to a child
      * {@link RealmObject} or a child {@link RealmList}.
      */
+    @Deprecated
     public RealmResults<E> findAllSorted(String fieldName1, Sort sortOrder1,
             String fieldName2, Sort sortOrder2) {
         return findAllSorted(new String[] {fieldName1, fieldName2}, new Sort[] {sortOrder1, sortOrder2});
@@ -1967,6 +2100,7 @@ public class RealmQuery<E> {
      * @throws java.lang.IllegalArgumentException if a field name does not exist or it belongs to a child
      * {@link RealmObject} or a child {@link RealmList}.
      */
+    @Deprecated
     public RealmResults<E> findAllSortedAsync(String fieldName1, Sort sortOrder1,
             String fieldName2, Sort sortOrder2) {
         return findAllSortedAsync(new String[] {fieldName1, fieldName2}, new Sort[] {sortOrder1, sortOrder2});